Implementacao fase final com sucesso e outros
Implementacao fase final com sucesso; mouse desabilitado; pomo de ouro mexe certo e sempre; e outros pequenos bugs; AINDA FALTA ARRUMAR FEITICO SAINDO NA TRANSICAO

+++ feitico.h
@@ -5,7 +5,7 @@
 #include <SFML/Window.hpp>
 #include <iostream>
 #include <math.h>
-#include "ResourcePath.hpp"
+ 
 
 using namespace std;
 
@@ -31,7 +31,7 @@ public:
 	virtual float get_x();
 	virtual float get_y();
 
-	static float dir;
+	static double dir;
 	static bool lancado;
 
 private:
